ajout des tri

+++ cycle5/Cycle5/src/cycle5/IndiceComparaison.java
@@ -9,15 +9,23 @@ package cycle5;
  * @author qartigala
  */
 public class IndiceComparaison {
-       
-    int indice;
-    
-    int comparaison;
-
-    public IndiceComparaison(int unIndice, int uneComparaison) {
+    /**
+     * Indice d'un élément du tableau
+     */  
+    double indice;
+    /**
+     * le nombre de comparaison
+     */
+    double comparaison;
+    /**
+     * Le constructeur de l'association des deux valeurs
+     * @param unIndice l'indice de l'élément du tableau
+     * @param uneComparaison le nombre de comparaison effecuter
+     */
+    public IndiceComparaison(double unIndice, double uneComparaison) {
         this.indice=unIndice;
         this.comparaison=uneComparaison;
     }
-    
+   
     
 }
